+++ tests/framework/bindings/BuiltinBinderTest.php
@@ -7,12 +7,8 @@
 
 namespace yiiunit\framework\bindings;
 
-use DateTime;
-use DateTimeImmutable;
-use ReflectionProperty;
-use yii\base\Controller;
+
 use yii\base\InlineAction;
-use yii\bindings\BindingProperty;
 use yii\bindings\ActionParameterBinder;
 use yii\bindings\binders\ActiveRecordBinder;
 use yii\bindings\binders\BuiltinTypeBinder;
@@ -22,74 +18,6 @@ use yii\bindings\binders\DataFilterBinder;
 use yii\bindings\BindingContext;
 use yii\bindings\ModelBinderInterface;
 use yiiunit\TestCase;
-
-class TypeReflector
-{
-    public array $array;
-    public int $int;
-    public float $float;
-    public bool $bool;
-    public string $string;
-    public ?int $nullable_int;
-    public ?float $nullable_float;
-    public ?bool $nullable_bool;
-    public ?string $nullable_string;
-    public DateTime $DateTime;
-    public DateTimeImmutable $DateTimeImmutable;
-
-    public static function getReflectionProperty($name)
-    {
-        $name = str_replace("?", "nullable_", $name);
-        $name = str_replace("\\", "_", $name);
-        return new ReflectionProperty(self::class, $name);
-    }
-
-    public static function getBindingTarget($name, $value)
-    {
-        return new BindingProperty(self::getReflectionProperty($name), $value);
-    }
-}
-
-class ComplexObject {
-    public int $int;
-    public float $float;
-    public bool $bool;
-}
-
-class TestController extends Controller
-{
-
-    public function actionParams(
-        $mixed,
-        int $int,
-        float $float,
-        bool $bool,
-        DateTime $dateTime,
-        DateTimeImmutable $dateTimeImmutable
-    )
-    {
-    }
-
-    public function actionNoType($value)
-    {
-    }
-
-    public function actionBuiltin(int $int, float $float, bool $bool)
-    {
-    }
-
-    public function actionBuiltinNullable(?int $int, ?float $float, ?bool $bool)
-    {
-    }
-
-    public function actionDateTime(DateTime $dateTime, DateTimeImmutable $dateTimeImmutable)
-    {
-    }
-
-    public function actionDateTimeNullable(?DateTime $dateTime, ?DateTimeImmutable $dateTimeImmutable)
-    {
-    }
-}
 
 class BuiltinBinderTest extends TestCase
 {
@@ -105,21 +33,6 @@ class BuiltinBinderTest extends TestCase
     private $modelBinder;
 
     /**
-     * @var ModelBinderInterface
-     */
-    private $dateTimeBinder;
-
-    /**
-     * @var ModelBinderInterface
-     */
-    private $containerBinder;
-
-    /**
-     * @var ModelBinderInterface
-     */
-    private $activeRecordBinder;
-
-    /**
      * @var BindingContext
      */
     private $context = null;
@@ -128,11 +41,7 @@ class BuiltinBinderTest extends TestCase
     {
         parent::setUp();
         $this->parameterBinder = new ActionParameterBinder();
-        $this->builtInBinder = new BuiltinTypeBinder();
-        $this->dateTimeBinder = new DateTimeBinder();
-        $this->containerBinder = new ContainerBinder();
-        $this->activeRecordBinder = new ActiveRecordBinder();
-        $this->dataFilterBinder = new DataFilterBinder();
+        $this->modelBinder = new BuiltinTypeBinder();
 
         $this->mockWebApplication([
             'components' => [
@@ -197,7 +106,7 @@ class BuiltinBinderTest extends TestCase
      */
     public function testBuiltInBinder($typeName, $expected, $value)
     {
-        $binding = $this->builtInBinder->bindModel(TypeReflector::getBindingTarget($typeName, $value), $this->context);
+        $binding = $this->modelBinder->bindModel(TypeReflector::getBindingTarget($typeName, $value), $this->context);
 
         if ($expected) {
             $this->assertNotNull($binding);
